optimize code (#469)
